requirements.txt
chromadb
pypdf
python-dotenv
sentence_transformers
